+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Current (in progress)
 
 - Improve reuse api perfs by adding a mask on datasets [#3309](https://github.com/opendatateam/udata/pull/3309)
+- Private objects should return 404 by api [#3311](https://github.com/opendatateam/udata/pull/3311)
 
 ## 10.3.2 (2025-05-06)