+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Current (in progress)
 
 - Improve reuse api perfs by adding a mask on datasets [#3309](https://github.com/opendatateam/udata/pull/3309)
+- Add `featured` to dataset default mask [#3313](https://github.com/opendatateam/udata/pull/3313)
 
 ## 10.3.2 (2025-05-06)