+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Improve reuse api perfs by adding a mask on datasets [#3309](https://github.com/opendatateam/udata/pull/3309)
 - Allow returning full sub-objects (license, frequency, zones and granularity) for datasets APIv2 [#3310](https://github.com/opendatateam/udata/pull/3310)
+- Private objects should return 404 by api [#3311](https://github.com/opendatateam/udata/pull/3311)
 - Add `featured` to dataset default mask [#3313](https://github.com/opendatateam/udata/pull/3313)
 
 ## 10.3.2 (2025-05-06)