+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Current (in progress)
 
 - Improve reuse api perfs by adding a mask on datasets [#3309](https://github.com/opendatateam/udata/pull/3309)
-- Private objects should return 404 by api [#3311](https://github.com/opendatateam/udata/pull/3311)
+- Private objects should return 404 by api [#3311](https://github.com/opendatateam/udata/pull/3311) [#3316](https://github.com/opendatateam/udata/pull/3316)
 - Add `featured` to dataset default mask [#3313](https://github.com/opendatateam/udata/pull/3313)
 
 ## 10.3.2 (2025-05-06)